LayoutTests/fast/js/string-property-deletion-expected.txt

This page tests deletion of properties on a string object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS str.length is 3
PASS delete str.length is false
PASS delete str[0] is false
PASS delete str[1] is false
PASS delete str[2] is false
PASS delete str[3] is true
PASS delete str[-1] is true
PASS delete str[4294967294] is true
PASS delete str[4294967295] is true
PASS delete str[4294967296] is true
PASS delete str[0.0] is false
PASS delete str[0.1] is true
PASS delete str['0.0'] is true
PASS delete str.foo is true
PASS successfullyParsed is true

TEST COMPLETE
